Extract screw primitive code.
