+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/behaviorTree/actions/RDXArmAction.java
@@ -1,11 +1,9 @@
 package us.ihmc.rdx.behaviorTree.actions;
 
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g3d.Renderable;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Pool;
 import imgui.ImGui;
-import imgui.flag.ImGuiCol;
 import imgui.flag.ImGuiMouseButton;
 import imgui.type.ImInt;
 import org.apache.commons.lang3.mutable.MutableObject;
@@ -15,19 +13,15 @@ import us.ihmc.behaviors.behaviorTree.action.actions.AbilityHandActionState;
 import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionDefinition;
 import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionState;
 import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionTaskspaceTrajectoryMode;
-import us.ihmc.commons.lists.RecyclingArrayList;
 import us.ihmc.commons.thread.Throttler;
-import us.ihmc.commons.time.Stopwatch;
 import us.ihmc.communication.crdt.CRDTBidirectionalRigidBodyTransform;
 import us.ihmc.communication.crdt.CRDTDetachableReferenceFrame;
-import us.ihmc.euclid.Axis3D;
 import us.ihmc.euclid.referenceFrame.FramePose3D;
 import us.ihmc.euclid.referenceFrame.ReferenceFrame;
 import us.ihmc.euclid.transform.interfaces.RigidBodyTransformReadOnly;
 import us.ihmc.handsros2.abilityHand.AbilityHandControlMode;
 import us.ihmc.handsros2.abilityHand.AbilityHandGrip;
 import us.ihmc.handsros2.abilityHand.AbilityHandModel.AbilityHandJointName;
-import us.ihmc.humanoidRobotics.frames.HumanoidReferenceFrames;
 import us.ihmc.mecano.frames.MovingReferenceFrame;
 import us.ihmc.mecano.multiBodySystem.RevoluteJoint;
 import us.ihmc.mecano.multiBodySystem.interfaces.MultiBodySystemBasics;
@@ -38,16 +32,16 @@ import us.ihmc.rdx.behaviorTree.RDXCRDTTools;
 import us.ihmc.rdx.imgui.*;
 import us.ihmc.rdx.input.ImGui3DViewInput;
 import us.ihmc.rdx.input.ImGui3DViewPickResult;
-import us.ihmc.rdx.mesh.RDXDashedLineMesh;
 import us.ihmc.rdx.simulation.scs2.RDXRigidBody;
+import us.ihmc.rdx.ui.RDX3DPanel;
 import us.ihmc.rdx.ui.RDX3DPanelTooltip;
 import us.ihmc.rdx.ui.affordances.RDXInteractableHighlightModel;
 import us.ihmc.rdx.ui.affordances.RDXInteractableTools;
 import us.ihmc.rdx.ui.gizmo.RDXSelectablePose3DGizmo;
 import us.ihmc.rdx.ui.graphics.RDXArmMultiBodyGraphic;
-import us.ihmc.rdx.ui.graphics.RDXTrajectoryGraphic;
 import us.ihmc.rdx.ui.teleoperation.RDXIKSolverColors;
 import us.ihmc.rdx.ui.widgets.ImGuiArmIconWidget;
+import us.ihmc.avatar.drcRobot.ROS2SyncedRobotModel;
 import us.ihmc.robotModels.FullHumanoidRobotModel;
 import us.ihmc.robotics.EuclidCoreMissingTools;
 import us.ihmc.robotics.MultiBodySystemMissingTools;
@@ -70,10 +64,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
    private final ImGuiArmIconWidget armIconWidget = new ImGuiArmIconWidget();
    /** Gizmo is control frame */
    private final RDXSelectablePose3DGizmo poseGizmo;
-   private final RDXSelectablePose3DGizmo screwAxisGizmo;
-   private final RDXDashedLineMesh screwAxisGraphic = new RDXDashedLineMesh(Color.WHITE, Axis3D.X, 0.04);
-   private final RDXTrajectoryGraphic trajectoryGraphic = new RDXTrajectoryGraphic();
-   private final RecyclingArrayList<FramePose3D> trajectoryPoses = new RecyclingArrayList<>(FramePose3D::new);
+   private final RDXScrewPrimitive screwPrimitive;
    private final SideDependentList<RigidBodyTransformReadOnly> handGraphicToControlFrameTransforms = new SideDependentList<>();
    private final MutableReferenceFrame graphicFrame = new MutableReferenceFrame();
    private final MutableReferenceFrame collisionShapeFrame = new MutableReferenceFrame();
@@ -89,25 +80,19 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
    private final SideDependentList<ArmJointName[]> jointNames = new SideDependentList<>();
    private final SideDependentList<double[]> jointLowerLimits = new SideDependentList<>();
    private final SideDependentList<double[]> jointUpperLimits = new SideDependentList<>();
+   private final ImGuiLabelledWidgetAligner widgetAligner = new ImGuiLabelledWidgetAligner();
    private final ImGuiSliderDoubleWrapper linearPositionWeightWidget;
    private final ImGuiSliderDoubleWrapper angularPositionWeightWidget;
    private final ImGuiSliderDoubleWrapper jointspaceWeightWidget;
    private final ImBooleanWrapper holdPoseInWorldLaterWrapper;
    private final ImBooleanWrapper jointSpaceControlWrapper;
    private final ImBooleanWrapper definedInJointspaceWrapper;
-   private final ImGuiSliderDoubleWrapper translationWidget;
-   private final ImGuiSliderDoubleWrapper rotationWidget;
-   private final ImGuiSliderDoubleWrapper maxLinearVelocityWidget;
-   private final ImGuiSliderDoubleWrapper maxAngularVelocityWidget;
-   private final ImGuiSliderDoubleWrapper previewTimeWidget;
    private final ImDoubleWrapper positionErrorToleranceInput;
    private final ImDoubleWrapper orientationErrorToleranceDegreesInput;
    private final SideDependentList<RDXArmMultiBodyGraphic> armMultiBodyGraphics = new SideDependentList<>();
    private final SideDependentList<RDXRigidBody> abilityHands = new SideDependentList<>();
    private final double[] fingerPositions = new double[6];
    private boolean showAbilityHand = false;
-   private boolean playbackPreview = false;
-   private final Stopwatch playbackStopwatch = new Stopwatch();
    private final Throttler lowQualityRenderThrottler = new Throttler();
    private final RDX3DPanelTooltip tooltip;
 
@@ -117,8 +102,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
 
       poseGizmo = new RDXSelectablePose3DGizmo();
       poseGizmo.create(panel3D);
-      screwAxisGizmo = new RDXSelectablePose3DGizmo();
-      screwAxisGizmo.create(panel3D);
+      screwPrimitive = new RDXScrewPrimitive(this);
 
       trajectoryDurationWidget = new ImDoubleWrapper(definition::getTrajectoryDuration,
                                                      definition::setTrajectoryDuration,
@@ -218,7 +202,6 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          if (ImGui.radioButton(labels.get("Jointspace Only"), imBoolean.get()))
             imBoolean.set(true);
       });
-      ImGuiLabelledWidgetAligner widgetAligner = new ImGuiLabelledWidgetAligner();
       linearPositionWeightWidget = new ImGuiSliderDoubleWrapper("Linear Position Weight", "%.2f", 0.0, 100.0,
                                                                 definition::getLinearPositionWeight,
                                                                 definition::setLinearPositionWeight);
@@ -234,25 +217,6 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
                                                             definition::setJointspaceWeight);
       jointspaceWeightWidget.addButton("Use Default Weights", () -> definition.setJointspaceWeight(-1.0));
       jointspaceWeightWidget.addWidgetAligner(widgetAligner);
-      translationWidget = new ImGuiSliderDoubleWrapper("Translation", "%.2f", -0.4, 0.4, definition::getTranslation, definition::setTranslation);
-      translationWidget.addWidgetAligner(widgetAligner);
-      rotationWidget = new ImGuiSliderDoubleWrapper("Rotation", "%.2f", -2.0 * Math.PI, 2.0 * Math.PI,
-                                                    definition::getRotation,
-                                                    definition::setRotation);
-      rotationWidget.addWidgetAligner(widgetAligner);
-      maxLinearVelocityWidget = new ImGuiSliderDoubleWrapper("Max Linear Velocity", "%.2f", 0.05, 1.0,
-                                                             definition::getMaxLinearVelocity,
-                                                             definition::setMaxLinearVelocity);
-      maxLinearVelocityWidget.addWidgetAligner(widgetAligner);
-      maxAngularVelocityWidget = new ImGuiSliderDoubleWrapper("Max Angular Velocity", "%.2f", 0.1, Math.PI,
-                                                              definition::getMaxAngularVelocity,
-                                                              definition::setMaxAngularVelocity);
-      maxAngularVelocityWidget.addWidgetAligner(widgetAligner);
-      previewTimeWidget = new ImGuiSliderDoubleWrapper("Preview Time", "%.2f", 0.0, 1.0,
-                                                       state.getPreviewRequestedTime()::getValue,
-                                                       state.getPreviewRequestedTime()::setValue);
-      previewTimeWidget.addButton("Play", this::togglePlayPausePreview);
-      previewTimeWidget.addWidgetAligner(widgetAligner);
       positionErrorToleranceInput = new ImDoubleWrapper(() ->
                                                         {
                                                            if (definition.getDefinedInJointspace())
@@ -320,18 +284,6 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       panel3D.addImGuiOverlayAddition(this::render3DPanelImGuiOverlays);
    }
 
-   private boolean isSinglePoseTaskspace()
-   {
-      return !definition.getDefinedInJointspace()
-          && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE;
-   }
-
-   private boolean isScrewTaskspace()
-   {
-      return !definition.getDefinedInJointspace()
-          && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE;
-   }
-
    @Override
    public void update()
    {
@@ -344,7 +296,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       if (definition.getDefinedInJointspace())
       {
          poseGizmo.setSelected(false);
-         screwAxisGizmo.setSelected(false);
+         screwPrimitive.deselectGizmos();
 
          PresetArmConfiguration preset = definition.getPreset();
          currentConfiguration.set(preset == null ? 0 : preset.ordinal() + 1);
@@ -358,6 +310,8 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       }
       else if (isSinglePoseTaskspace() && state.getPalmFrame().isChildOfWorld())
       {
+         screwPrimitive.deselectGizmos();
+
          if (poseGizmo.getPoseGizmo().getGizmoFrame() != state.getPalmFrame().getReferenceFrame())
          {
             poseGizmo.getPoseGizmo().setGizmoFrame(state.getPalmFrame().getReferenceFrame());
@@ -375,32 +329,10 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          else
             highlightModels.get(definition.getSide()).setTransparency(0.5);
       }
-      else if (isScrewTaskspace() && state.getScrewFrame().isChildOfWorld())
+      else if (isScrewTaskspace())
       {
-         if (screwAxisGizmo.getPoseGizmo().getGizmoFrame() != state.getScrewFrame().getReferenceFrame())
-            screwAxisGizmo.getPoseGizmo().setGizmoFrame(state.getScrewFrame().getReferenceFrame());
-
-         if (!getSelected())
-            screwAxisGizmo.setSelected(false);
-
-         RDXCRDTTools.syncGizmoWithBidirectionalField(screwAxisGizmo.getPoseGizmo(), definition.getScrewAxisPoseInObjectFrame(), definition);
-
-         double screwAxisLineWidth = 0.005;
-         screwAxisGraphic.update(screwAxisGizmo.getPoseGizmo().getPose(), screwAxisLineWidth, 1.0);
-
-         double trajectoryLineWidth = 0.01;
-         trajectoryPoses.clear();
-         for (int i = 0; i < state.getPreviewTrajectory().getSize(); i++)
-            trajectoryPoses.add().set(state.getPreviewTrajectory().getValueReadOnly(i));
-         trajectoryGraphic.update(trajectoryLineWidth, trajectoryPoses);
-
-         if (playbackPreview)
-         {
-            double requestedTime = state.getPreviewRequestedTime().getValue();
-            requestedTime += playbackStopwatch.lap() / state.getPreviewTrajectoryDuration().getValue();
-            requestedTime %= 1.0;
-            state.getPreviewRequestedTime().setValue(requestedTime);
-         }
+         poseGizmo.setSelected(false);
+         screwPrimitive.update();
       }
    }
 
@@ -410,11 +342,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
 
       if (isScrewTaskspace())
       {
-         armMultiBodyGraphic.getFloatingJoint().getJointPose().set(syncedRobot.getFramePoseReadOnly(HumanoidReferenceFrames::getChestFrame));
-         for (int i = 0; i < armMultiBodyGraphic.getJoints().length; i++)
-            armMultiBodyGraphic.getJoints()[i].setQ(state.getScrewPreviewJointAngles().getValueReadOnly(i));
-         armMultiBodyGraphic.updateAfterModifyingConfiguration();
-         armMultiBodyGraphic.setColor(RDXIKSolverColors.getColor(state.getPreviewSolutionQuality().getValue()));
+         screwPrimitive.visualizeIK(armMultiBodyGraphic);
       }
       else
       {
@@ -582,50 +510,15 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       }
       else if (isScrewTaskspace())
       {
-         ImGui.checkbox(labels.get("Adjust Screw Axis Pose"), screwAxisGizmo.getSelected());
-         parentFrameComboBox.render();
-         int size = state.getPreviewTrajectory().getSize();
-         int limit = ArmActionState.TRAJECTORY_SIZE_LIMIT;
-         if (size == limit)
-            ImGui.pushStyleColor(ImGuiCol.Text, ImGuiTools.RED);
-         ImGui.text("Trajectory points: %d/%d  Duration: %.1f s  Velocity %.2f m/s  %.2f %s/s"
-                          .formatted(size,
-                                     limit,
-                                     state.getPreviewTrajectoryDuration().getValue(),
-                                     state.getPreviewTrajectoryLinearVelocity().getValue(),
-                                     state.getPreviewTrajectoryAngularVelocity().getValue(),
-                                     EuclidCoreMissingTools.DEGREE_SYMBOL));
-         if (size == limit)
-            ImGui.popStyleColor();
-         translationWidget.renderImGuiWidget();
-         rotationWidget.renderImGuiWidget();
-         maxLinearVelocityWidget.renderImGuiWidget();
-         maxAngularVelocityWidget.renderImGuiWidget();
-         linearPositionWeightWidget.renderImGuiWidget();
-         angularPositionWeightWidget.renderImGuiWidget();
-         jointspaceWeightWidget.renderImGuiWidget();
-         ImGui.pushItemWidth(ImGui.getFontSize() * 10.0f);
-         positionErrorToleranceInput.renderImGuiWidget();
-         orientationErrorToleranceDegreesInput.renderImGuiWidget();
-         ImGui.popItemWidth();
-         previewTimeWidget.renderImGuiWidget();
+         screwPrimitive.renderImGuiWidgets();
       }
-   }
-
-   private void togglePlayPausePreview()
-   {
-      playbackPreview = !playbackPreview;
-      previewTimeWidget.addButton(playbackPreview ? "Pause" : "Play", this::togglePlayPausePreview);
-
-      if (playbackPreview)
-         playbackStopwatch.reset();
    }
 
    @Override
    public void deselectGizmos()
    {
       poseGizmo.setSelected(false);
-      screwAxisGizmo.setSelected(false);
+      screwPrimitive.deselectGizmos();
    }
 
    public void render3DPanelImGuiOverlays()
@@ -654,9 +547,9 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          if (pickResult.getPickCollisionWasAddedSinceReset())
             input.addPickResult(pickResult);
       }
-      else if (isScrewTaskspace() && state.getScrewFrame().isChildOfWorld())
+      else if (isScrewTaskspace())
       {
-         screwAxisGizmo.calculate3DViewPick(input);
+         screwPrimitive.calculate3DViewPick(input);
       }
    }
 
@@ -677,9 +570,9 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          poseGizmo.process3DViewInput(input, isMouseHovering);
          tooltip.setInput(input);
       }
-      else if (isScrewTaskspace() && state.getScrewFrame().isChildOfWorld())
+      else if (isScrewTaskspace())
       {
-         screwAxisGizmo.process3DViewInput(input);
+         screwPrimitive.process3DViewInput(input);
       }
    }
 
@@ -695,11 +588,9 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
             highlightModels.get(definition.getSide()).getRenderables(renderables, pool);
          poseGizmo.getVirtualRenderables(renderables, pool);
       }
-      else if (isScrewTaskspace() && state.getScrewFrame().isChildOfWorld())
+      else if (isScrewTaskspace())
       {
-         screwAxisGizmo.getVirtualRenderables(renderables, pool);
-         screwAxisGraphic.getRenderables(renderables, pool);
-         trajectoryGraphic.getRenderables(renderables, pool);
+         screwPrimitive.getRenderables(renderables, pool);
       }
 
       if (state.getIsNextForExecution() || state.getIsExecuting())
@@ -714,12 +605,74 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
    public String getLeafTypeTitle()
    {
       if (isScrewTaskspace())
-         return definition.getSide().getPascalCaseName() + " Screw Primitive";
+         return screwPrimitive.getLeafTypeTitle();
       return definition.getSide().getPascalCaseName() + " Arm";
+   }
+
+   private boolean isSinglePoseTaskspace()
+   {
+      return !definition.getDefinedInJointspace()
+             && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE;
+   }
+
+   private boolean isScrewTaskspace()
+   {
+      return !definition.getDefinedInJointspace()
+             && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE;
    }
 
    public ReferenceFrame getReferenceFrame()
    {
       return poseGizmo.getPoseGizmo().getGizmoFrame();
+   }
+
+   RDX3DPanel getPanel3D()
+   {
+      return panel3D;
+   }
+
+   ROS2SyncedRobotModel getSyncedRobot()
+   {
+      return syncedRobot;
+   }
+
+   ImGuiUniqueLabelMap getLabels()
+   {
+      return labels;
+   }
+
+   ImGuiLabelledWidgetAligner getWidgetAligner()
+   {
+      return widgetAligner;
+   }
+
+   ImGuiReferenceFrameLibraryCombo getParentFrameComboBox()
+   {
+      return parentFrameComboBox;
+   }
+
+   ImGuiSliderDoubleWrapper getLinearPositionWeightWidget()
+   {
+      return linearPositionWeightWidget;
+   }
+
+   ImGuiSliderDoubleWrapper getAngularPositionWeightWidget()
+   {
+      return angularPositionWeightWidget;
+   }
+
+   ImGuiSliderDoubleWrapper getJointspaceWeightWidget()
+   {
+      return jointspaceWeightWidget;
+   }
+
+   ImDoubleWrapper getPositionErrorToleranceInput()
+   {
+      return positionErrorToleranceInput;
+   }
+
+   ImDoubleWrapper getOrientationErrorToleranceDegreesInput()
+   {
+      return orientationErrorToleranceDegreesInput;
    }
 }
