+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/action/actions/ArmActionState.java
@@ -15,11 +15,10 @@ import static us.ihmc.behaviors.behaviorTree.action.actions.ArmActionDefinition.
 
 public class ArmActionState extends ActionNodeState<ArmActionDefinition>
 {
-   /** This limit is defined in the .msg file and limited to the size in the SE3TrajectoryMessage. */
-   public static final int TRAJECTORY_SIZE_LIMIT = new ArmActionStateMessage().getPreviewTrajectory().getCurrentCapacity();
+   public static final int TRAJECTORY_SIZE_LIMIT = ScrewPrimitiveState.TRAJECTORY_SIZE_LIMIT;
 
    private final CRDTDetachableReferenceFrame palmFrame;
-   private final DetachableReferenceFrame screwFrame;
+   private final ScrewPrimitiveState screwPrimitive;
    /**
     * This is the estimated goal chest frame as the robot executes a potential whole body action.
     * This is used to compute joint angles that achieve the desired and previewed end pose
@@ -31,13 +30,6 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
    private final CRDTStatusVector3D torque;
    private final CRDTStatusDoubleArray previewJointAngles;
    private final CRDTStatusDouble solutionQuality;
-   private final CRDTStatusPoseList previewTrajectory;
-   private final CRDTStatusDouble previewTrajectoryDuration;
-   private final CRDTStatusDouble previewTrajectoryLinearVelocity;
-   private final CRDTStatusDouble previewTrajectoryAngularVelocity;
-   private final CRDTStatusDouble previewRequestedTime;
-   private final CRDTStatusDoubleArray screwPreviewJointAngles;
-   private final CRDTStatusDouble previewSolutionQuality;
    private final SideDependentList<Integer> numberOfJoints = new SideDependentList<>();
 
    public ArmActionState(long id, BehaviorTreeRootNodeState rootNode)
@@ -47,7 +39,7 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       palmFrame = new CRDTDetachableReferenceFrame(scene::findFrameByName,
                                                    definition.getCRDTPalmParentFrameName(),
                                                    definition.getPalmTransformToParent());
-      screwFrame = new DetachableReferenceFrame(scene::findFrameByName, definition.getScrewAxisPoseInObjectFrame().getValueReadOnly());
+      screwPrimitive = new ScrewPrimitiveState(scene, crdtInfo, definition.getScrewPrimitive());
       goalChestToWorldTransform = new CRDTStatusRigidBodyTransform(ROS2ActorDesignation.ROBOT, crdtInfo);
       goalChestFrame = ReferenceFrameMissingTools.constructFrameWithChangingTransformToParent(ReferenceFrame.getWorldFrame(),
                                                                                               goalChestToWorldTransform.getValueReadOnly());
@@ -55,13 +47,6 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       torque = new CRDTStatusVector3D(ROS2ActorDesignation.ROBOT, crdtInfo);
       previewJointAngles = new CRDTStatusDoubleArray(ROS2ActorDesignation.ROBOT, crdtInfo, MAX_NUMBER_OF_JOINTS);
       solutionQuality = new CRDTStatusDouble(ROS2ActorDesignation.ROBOT, crdtInfo, Double.NaN);
-      previewTrajectory = new CRDTStatusPoseList(ROS2ActorDesignation.ROBOT, crdtInfo);
-      previewTrajectoryDuration = new CRDTStatusDouble(ROS2ActorDesignation.ROBOT, crdtInfo, -1.0);
-      previewTrajectoryLinearVelocity = new CRDTStatusDouble(ROS2ActorDesignation.ROBOT, crdtInfo, -1.0);
-      previewTrajectoryAngularVelocity = new CRDTStatusDouble(ROS2ActorDesignation.ROBOT, crdtInfo, -1.0);
-      previewRequestedTime = new CRDTStatusDouble(ROS2ActorDesignation.OPERATOR, crdtInfo, 1.0);
-      screwPreviewJointAngles = new CRDTStatusDoubleArray(ROS2ActorDesignation.ROBOT, crdtInfo, MAX_NUMBER_OF_JOINTS);
-      previewSolutionQuality = new CRDTStatusDouble(ROS2ActorDesignation.ROBOT, crdtInfo, 0.0);
 
       for (RobotSide side : RobotSide.values)
          numberOfJoints.put(side, robotModel.getJointMap().getArmJointNamesAsStrings(side).size());
@@ -71,7 +56,7 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
    public void update()
    {
       palmFrame.update();
-      screwFrame.update(definition.getPalmParentFrameName());
+      screwPrimitive.update(definition.getPalmParentFrameName());
    }
 
    @Override
@@ -83,13 +68,7 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       hasStatus |= torque.pollHasStatus();
       hasStatus |= previewJointAngles.pollHasStatus();
       hasStatus |= solutionQuality.pollHasStatus();
-      hasStatus |= previewTrajectory.pollHasStatus();
-      hasStatus |= previewTrajectoryDuration.pollHasStatus();
-      hasStatus |= previewTrajectoryLinearVelocity.pollHasStatus();
-      hasStatus |= previewTrajectoryAngularVelocity.pollHasStatus();
-      hasStatus |= previewRequestedTime.pollHasStatus();
-      hasStatus |= screwPreviewJointAngles.pollHasStatus();
-      hasStatus |= previewSolutionQuality.pollHasStatus();
+      hasStatus |= screwPrimitive.pollHasStatus();
       return hasStatus;
    }
 
@@ -104,13 +83,7 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       torque.toMessage(message.getTorque());
       previewJointAngles.toMessage(message.getJointAngles());
       message.setSolutionQuality(solutionQuality.toMessage());
-      previewTrajectory.toMessage(message.getPreviewTrajectory());
-      message.setPreviewTrajectoryDuration(previewTrajectoryDuration.toMessage());
-      message.setPreviewTrajectoryLinearVelocity(previewTrajectoryLinearVelocity.toMessage());
-      message.setPreviewTrajectoryAngularVelocity(previewTrajectoryAngularVelocity.toMessage());
-      message.setPreviewRequestedTime(previewRequestedTime.toMessage());
-      screwPreviewJointAngles.toMessage(message.getPreviewJointAngles());
-      message.setPreviewSolutionQuality(previewSolutionQuality.toMessage());
+      screwPrimitive.toMessage(message);
    }
 
    public void fromMessage(ArmActionStateMessage message)
@@ -125,13 +98,7 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       solutionQuality.fromMessage(message.getSolutionQuality());
       goalChestToWorldTransform.fromMessage(message.getGoalChestTransformToWorld());
       goalChestFrame.update();
-      previewTrajectory.fromMessage(message.getPreviewTrajectory());
-      previewTrajectoryDuration.fromMessage(message.getPreviewTrajectoryDuration());
-      previewTrajectoryLinearVelocity.fromMessage(message.getPreviewTrajectoryLinearVelocity());
-      previewTrajectoryAngularVelocity.fromMessage(message.getPreviewTrajectoryAngularVelocity());
-      previewRequestedTime.fromMessage(message.getPreviewRequestedTime());
-      screwPreviewJointAngles.fromMessage(message.getPreviewJointAngles());
-      previewSolutionQuality.fromMessage(message.getPreviewSolutionQuality());
+      screwPrimitive.fromMessage(message);
    }
 
    public CRDTDetachableReferenceFrame getPalmFrame()
@@ -139,9 +106,14 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       return palmFrame;
    }
 
+   public ScrewPrimitiveState getScrewPrimitive()
+   {
+      return screwPrimitive;
+   }
+
    public DetachableReferenceFrame getScrewFrame()
    {
-      return screwFrame;
+      return screwPrimitive.getScrewFrame();
    }
 
    public CRDTStatusRigidBodyTransform getGoalChestToWorldTransform()
@@ -171,37 +143,37 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
 
    public CRDTStatusPoseList getPreviewTrajectory()
    {
-      return previewTrajectory;
+      return screwPrimitive.getPreviewTrajectory();
    }
 
    public CRDTStatusDouble getPreviewTrajectoryDuration()
    {
-      return previewTrajectoryDuration;
+      return screwPrimitive.getPreviewTrajectoryDuration();
    }
 
    public CRDTStatusDouble getPreviewTrajectoryLinearVelocity()
    {
-      return previewTrajectoryLinearVelocity;
+      return screwPrimitive.getPreviewTrajectoryLinearVelocity();
    }
 
    public CRDTStatusDouble getPreviewTrajectoryAngularVelocity()
    {
-      return previewTrajectoryAngularVelocity;
+      return screwPrimitive.getPreviewTrajectoryAngularVelocity();
    }
 
    public CRDTStatusDouble getPreviewRequestedTime()
    {
-      return previewRequestedTime;
+      return screwPrimitive.getPreviewRequestedTime();
    }
 
    public CRDTStatusDoubleArray getScrewPreviewJointAngles()
    {
-      return screwPreviewJointAngles;
+      return screwPrimitive.getScrewPreviewJointAngles();
    }
 
    public CRDTStatusDouble getPreviewSolutionQuality()
    {
-      return previewSolutionQuality;
+      return screwPrimitive.getPreviewSolutionQuality();
    }
 
    public int getNumberOfJoints()
